Add FindVisualStudio
Using Visual Studio's CppUnitTest framework requires an include and lib directory to be
added to the project paths.  By default, Visual Studio uses an internal VCInstallDir variable
which we can't use from CMake because CMake uses absolute paths.  The FindVisualStudio
allows to find the absolute paths of the desired directories.